+++ packages/article/article.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Platform, Text, View, ListView } from "react-native";
+import { Platform, Text, View } from "react-native";
 import PropTypes from "prop-types";
 import get from "lodash.get";
 import { renderTrees } from "@times-components/markup";
@@ -8,6 +8,7 @@ import ArticleImage from "@times-components/article-image";
 import { AdComposer } from "@times-components/ad";
 import { withTrackingContext } from "@times-components/tracking";
 
+import ArticleContent from "./article-content";
 import ArticleError from "./article-error";
 import ArticleLoading from "./article-loading";
 
@@ -16,7 +17,6 @@ import styles from "./styles/body";
 import ArticleHeader from "./article-header";
 import ArticleMeta from "./article-meta";
 
-const ds = new ListView.DataSource({ rowHasChanged: (r1, r2) => r1 !== r2 });
 const listViewPageSize = 1;
 const listViewSize = 10;
 const listViewScrollRenderAheadDistance = 10;
@@ -28,19 +28,18 @@ const withAdComposer = (children, section = "article") => (
 class ArticlePage extends React.Component {
   static renderRow(rowData) {
     if (rowData.type === "leadAsset") {
+      const [ratioWidth, ratioHeight] = rowData.data.crop.ratio.split(":");
+      const aspectRatio = ratioWidth / ratioHeight;
       return (
-        <View
-          testID="leadAsset"
-          accessibilityLabel="leadAsset"
-          style={styles.leadAsset}
-        >
-          <Image uri={rowData.data.crop.url} aspectRatio={16 / 9} />
+        <View testID="leadAsset" key={rowData.type} style={styles.leadAsset}>
+          <Image uri={rowData.data.crop.url} aspectRatio={aspectRatio} />
         </View>
       );
     } else if (rowData.type === "header") {
       const { headline, flags, standfirst, label } = rowData.data;
       return (
         <ArticleHeader
+          key={rowData.type}
           headline={headline}
           flags={flags}
           standfirst={standfirst}
@@ -51,6 +50,7 @@ class ArticlePage extends React.Component {
       const { byline, publishedTime, publicationName } = rowData.data;
       return (
         <ArticleMeta
+          key={rowData.type}
           byline={byline}
           publishedTime={publishedTime}
           publicationName={publicationName}
@@ -58,7 +58,7 @@ class ArticlePage extends React.Component {
       );
     } else if (rowData.type === "articleBodyRow") {
       return (
-        <View>
+        <View key={rowData.type + rowData.index}>
           {renderTrees([rowData.data], {
             paragraph(key, attributes, children) {
               return (
@@ -101,11 +101,11 @@ class ArticlePage extends React.Component {
 
     if (props.article && !props.isLoading && !props.error) {
       this.state = {
-        dataSource: ds.cloneWithRows(listViewDataHelper(props.article))
+        dataSource: listViewDataHelper(props.article)
       };
     } else {
       this.state = {
-        dataSource: ds.cloneWithRows({})
+        dataSource: {}
       };
     }
   }
@@ -113,7 +113,7 @@ class ArticlePage extends React.Component {
   componentWillReceiveProps(nextProps) {
     if (!nextProps.isLoading && !nextProps.error) {
       this.setState({
-        dataSource: ds.cloneWithRows(listViewDataHelper(nextProps.article))
+        dataSource: listViewDataHelper(nextProps.article)
       });
     }
   }
@@ -128,17 +128,13 @@ class ArticlePage extends React.Component {
     if (isLoading) {
       return <ArticleLoading />;
     }
-
     const ArticleListView = (
-      <ListView
-        testID="listView"
-        accessibilityLabel="listView"
-        dataSource={this.state.dataSource}
+      <ArticleContent
+        data={this.state.dataSource}
         renderRow={ArticlePage.renderRow}
         initialListSize={listViewSize}
         scrollRenderAheadDistance={listViewScrollRenderAheadDistance}
         pageSize={listViewPageSize}
-        enableEmptySections
       />
     );