+++ packages/author-profile/__tests__/author-profile-helper.js
@@ -11,12 +11,14 @@ import { MockedProvider, mockNetworkInterface } from "react-apollo/test-utils";
 import { addTypenameToDocument } from "apollo-client";
 import { query as authorProfileQuery } from "@times-components/provider/author-profile-provider";
 import { query as articleListQuery } from "@times-components/provider/article-list-provider";
+import set from "lodash.set";
+import cloneDeep from "lodash.clonedeep";
 import AuthorProfile from "../author-profile";
 import AuthorProfileItem from "../author-profile-item";
 import AuthorHead from "../author-profile-author-head";
 import AuthorProfileItemSeparator from "../author-profile-item-separator";
 import authorProfileFixture from "../fixtures/author-profile.json";
-import articleListFixture from "../fixtures/article-list.json";
+import pagedResult from "./paged-result";
 
 Enzyme.configure({ adapter: new React16Adapter() });
 
@@ -26,23 +28,6 @@ const props = {
   onArticlePress: () => {},
   analyticsStream: () => {}
 };
-
-const pagedResult = (skip, first) => ({
-  data: {
-    author: {
-      ...articleListFixture.data.author,
-      articles: {
-        ...articleListFixture.data.author.articles,
-        list: articleListFixture.data.author.articles.list
-          .map(el => ({
-            ...el,
-            publishedTime: new Date(el.publishedTime)
-          }))
-          .slice(skip, skip + first)
-      }
-    }
-  }
-});
 
 const mocks = [
   {
@@ -155,8 +140,6 @@ export default AuthorProfileContent => {
           slug="fiona-hamilton"
           page={1}
           pageSize={10}
-          onTwitterLinkPress={() => {}}
-          onArticlePress={() => {}}
         />
       )
     );
@@ -174,7 +157,8 @@ export default AuthorProfileContent => {
         .map((number, id) => ({
           id,
           loading: true
-        }))
+        })),
+      imageRatio: 3 / 2
     };
 
     const component = renderer.create(<AuthorProfileContent {...p} />);
@@ -185,7 +169,8 @@ export default AuthorProfileContent => {
     const p = Object.assign({}, props, {
       slug: "fiona-hamilton",
       author: null,
-      isLoading: false
+      isLoading: false,
+      imageRatio: 16 / 9
     });
 
     const component = renderer.create(<AuthorProfileContent {...p} />);
@@ -253,8 +238,10 @@ export default AuthorProfileContent => {
         slug="fiona-hamilton"
         page={1}
         pageSize={3}
+        imageRatio={3 / 2}
         onTwitterLinkPress={() => {}}
         onArticlePress={() => {}}
+        onViewed={() => {}}
       />
     );
 
@@ -264,16 +251,36 @@ export default AuthorProfileContent => {
   it("renders profile content item component", () => {
     const item = pagedResult(0, 1).data.author.articles.list[0];
     const component = renderer.create(
-      <AuthorProfileItem {...item} onPress={() => {}} />
+      <AuthorProfileItem
+        {...item}
+        imageRatio={8 / 5}
+        imageSize={200}
+        onPress={() => {}}
+      />
     );
 
     expect(component).toMatchSnapshot();
   });
 
-  it("renders profile content item component with no image ", () => {
+  it("renders profile content item component with a specific image size", () => {
     const item = pagedResult(0, 1).data.author.articles.list[0];
     const component = renderer.create(
-      <AuthorProfileItem {...item} imageUri={null} onPress={() => {}} />
+      <AuthorProfileItem
+        {...item}
+        imageRatio={8 / 5}
+        imageSize={200}
+        onPress={() => {}}
+      />
+    );
+
+    expect(component).toMatchSnapshot();
+  });
+
+  it("renders profile content item component with no image", () => {
+    const item = cloneDeep(pagedResult(0, 1).data.author.articles.list[0]);
+    set(item, "leadAsset.crop.url", null);
+    const component = renderer.create(
+      <AuthorProfileItem {...item} imageRatio={20 / 3} onPress={() => {}} />
     );
 
     expect(component).toMatchSnapshot();
@@ -344,8 +351,6 @@ export default AuthorProfileContent => {
           slug="fiona-hamilton"
           page={1}
           pageSize={10}
-          onTwitterLinkPress={() => {}}
-          onArticlePress={() => {}}
           analyticsStream={stream}
         />
       )
@@ -378,8 +383,9 @@ export default AuthorProfileContent => {
       component: "AuthorProfileItem",
       action: "Pressed",
       attrs: {
-        articleId: "97c64f20-cb67-11e4-a202-50ac5def393a",
-        articleTitle: "British trio stopped on the way to join Isis"
+        articleId: "d98c257c-cb16-11e7-b529-95e3fc05f40f",
+        articleHeadline:
+          "Top medal for forces dog who took a bite out of the Taliban"
       }
     });
   });