+++ packages/author-profile/author-profile.js
@@ -8,6 +8,13 @@ import get from "lodash.get";
 import AuthorProfileError from "./author-profile-error";
 import AuthorProfileContent from "./author-profile-content";
 
+const ratioTextToFloat = s => {
+  const [w, h] = s.split(":");
+  const ratio = parseFloat(w) / parseFloat(h);
+
+  return !Number.isNaN(ratio) ? ratio : 1;
+};
+
 const ArticleListProviderWithPageState = withPageState(ArticleListProvider);
 const AuthorProfile = ({
   author,
@@ -39,29 +46,39 @@ const AuthorProfile = ({
         onPrev,
         page,
         pageSize,
-        isLoading: articlesLoading
-      }) => (
-        <AuthorProfileContent
-          isLoading={isLoading}
-          name={name}
-          biography={biography}
-          uri={uri}
-          jobTitle={jobTitle}
-          twitter={twitter}
-          onTwitterLinkPress={onTwitterLinkPress}
-          count={get(articles, "count", 0)}
-          onNext={onNext}
-          onPrev={onPrev}
-          page={page}
-          pageSize={pageSize}
-          articlesLoading={articlesLoading}
-          articles={get(data, "articles.list", []).map(article => ({
+        isLoading: articlesLoading,
+        variables: { imageRatio }
+      }) => {
+        const articlesWithPublishTime = get(data, "articles.list", []).map(
+          article => ({
             ...article,
+            author,
+            page,
+            pageSize,
             publishedTime: new Date(article.publishedTime)
-          }))}
-          onArticlePress={onArticlePress}
-        />
-      )}
+          })
+        );
+        return (
+          <AuthorProfileContent
+            isLoading={isLoading}
+            name={name}
+            biography={biography}
+            uri={uri}
+            jobTitle={jobTitle}
+            twitter={twitter}
+            onTwitterLinkPress={onTwitterLinkPress}
+            count={get(articles, "count", 0)}
+            onNext={onNext}
+            onPrev={onPrev}
+            page={page}
+            pageSize={pageSize}
+            imageRatio={ratioTextToFloat(imageRatio)}
+            articlesLoading={articlesLoading}
+            articles={articlesWithPublishTime}
+            onArticlePress={onArticlePress}
+          />
+        );
+      }}
     </ArticleListProviderWithPageState>
   );
 };
@@ -94,10 +111,10 @@ AuthorProfile.propTypes = {
 };
 
 export default withTrackingContext(AuthorProfile, {
-  trackingObject: "AuthorProfile",
   getAttrs: ({ author, page, pageSize } = {}) => ({
     authorName: author && author.name,
     page,
     pageSize
-  })
+  }),
+  trackingObject: "AuthorProfile"
 });