+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.8.13"></a>
+## [0.8.13](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.8.12...@times-components/article-summary@0.8.13) (2017-11-24)
+
+
+
+
 <a name="0.8.12"></a>
 ## [0.8.12](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.8.11...@times-components/article-summary@0.8.12) (2017-11-22)